fix: satisfy promote rollout ci gates
