+++ docs/changelog/entries/pr-742.json
@@ -1,0 +1,4 @@
+{
+  "prNumber": 742,
+  "body": "Der Staging-Promote kann gehärtete Migrations- und Bootstrap-One-shot-Jobs mit nachvollziehbarer Evidenz ausführen."
+}
