+++ scripts/ci/promote-deploy-gates.ts
@@ -1,12 +1,13 @@
 #!/usr/bin/env node
-import { appendFileSync } from 'node:fs';
 import { pathToFileURL } from 'node:url';
 import { parseBoolean, parseMode } from './promote-deploy-gates-cli.ts';
+import { evaluateEnvironmentRunGate, type PromoteEnvironment } from './promote-environment-gate.ts';
+import { emitPromoteDeployGateOutputs } from './promote-deploy-gate-output.ts';
 import { resolveChangedFiles } from './pr-scope.ts';
 import { resolveTraefikOnlyComposeFiles } from './traefik-compose-diff.ts';
 export type DeployGateMode = 'assert-none' | 'run';
 export type DeployGateKind = 'bootstrap' | 'migration';
-export type PromoteEnvironment = 'dev' | 'staging' | 'prod';
+export type { PromoteEnvironment } from './promote-environment-gate.ts';
 export type DeployGateResultType =
   'asserted-clean' | 'blocked-missing-executor' | 'blocked-risk' | 'blocked-safe-run-required';
 export interface DeployGateResult {
@@ -123,10 +124,11 @@ export const evaluateDeployGate = ({
       riskFiles,
     };
   }
-  if (environment === 'prod') {
+  const environmentGate = evaluateEnvironmentRunGate({ environment, label });
+  if (!environmentGate.ok) {
     return {
       kind,
-      message: `${label}-Gate blockiert: Production erlaubt One-shot-Jobs im Modus "run" noch nicht. Erforderlich sind Staging-Paritaet, Production-Freigabe, Backup-/Restore-Readiness und spezifische Postconditions.`,
+      message: environmentGate.message,
       mode,
       ok: false,
       result: 'blocked-safe-run-required',
@@ -134,23 +136,12 @@ export const evaluateDeployGate = ({
       riskFiles,
     };
   }
-  if (kind === 'bootstrap' || environment === 'dev' || environment === 'staging') {
-    return {
-      kind,
-      message: 'Bootstrap-Gate freigegeben: Der gehärtete One-shot-Executor wird im Promote-Workflow mit Exit-Code-Evidenz ausgeführt.',
-      mode,
-      ok: true,
-      result: 'asserted-clean',
-      riskDetected: riskFiles.length > 0,
-      riskFiles,
-    };
-  }
   return {
     kind,
-    message: `${label}-Gate blockiert: Ein Executor ist konfiguriert, aber im Promote-Workflow nicht mit gehärteter Exit-Code-/Log-Evidenz verdrahtet. Nutze den kanonischen Operator-Pfad statt Blindautomatisierung.`,
+    message: environmentGate.message,
     mode,
-    ok: false,
-    result: 'blocked-safe-run-required',
+    ok: true,
+    result: 'asserted-clean',
     riskDetected: riskFiles.length > 0,
     riskFiles,
   };
@@ -269,28 +260,6 @@ const resolveCliChangedFiles = ({
 }: Pick<CliOptions, 'base' | 'changedFiles' | 'head'>): string[] =>
   changedFiles ?? resolveChangedFiles(base, head);
 
-const emitEvaluationOutputs = (evaluation: PromoteDeployGateEvaluation): void => {
-  const combinedOk = evaluation.migration.ok && evaluation.bootstrap.ok ? 'true' : 'false';
-  const changedFilesSummary = evaluation.changedFiles.join(',');
-  const githubOutput = process.env.GITHUB_OUTPUT?.trim();
-  if (!githubOutput) {
-    return;
-  }
-  const lines = [
-    `changed_files=${changedFilesSummary}`,
-    `combined_ok=${combinedOk}`,
-    `migration_gate_ok=${String(evaluation.migration.ok)}`,
-    `migration_gate_result=${evaluation.migration.result}`,
-    `migration_gate_risk_files=${evaluation.migration.riskFiles.join(',') || 'none'}`,
-    `migration_gate_message=${evaluation.migration.message}`,
-    `bootstrap_gate_ok=${String(evaluation.bootstrap.ok)}`,
-    `bootstrap_gate_result=${evaluation.bootstrap.result}`,
-    `bootstrap_gate_risk_files=${evaluation.bootstrap.riskFiles.join(',') || 'none'}`,
-    `bootstrap_gate_message=${evaluation.bootstrap.message}`,
-  ];
-  appendFileSync(githubOutput, `${lines.join('\n')}\n`, 'utf8');
-};
-
 export const executePromoteDeployGates = async (
   args: readonly string[]
 ): Promise<{ exitCode: number; stderr: string; stdout: string }> => {
@@ -309,7 +278,7 @@ export const executePromoteDeployGates = async (
         : resolveTraefikOnlyComposeFiles(options.base, options.head, changedFiles),
     });
 
-    emitEvaluationOutputs(evaluation);
+    emitPromoteDeployGateOutputs(evaluation);
     return {
       exitCode: 0,
       stderr: '',
